worknest/src/app/backend/micro_approval/requirements.txt
Flask==3.0.1
pika==1.3.1  
google-api-python-client==2.92.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1